Run shellcheck on scripts
Shellcheck mostly flags potential pitfalls that may occur when paths have
unexpected spaces in them, so this is a lot of double-quoting fixes. I've
done this in order to make the scripts shellcheck-clean so that I can more
effectively use shellcheck on the additions I'm planning in the next
commits.